move(MemberService): member.application으로 옮김
